Ex4

+++ Ex3.java
@@ -6,14 +6,14 @@ public class Ex3 {
         System.out.println("Digite 10 numeros separados por espaço:");
         String valores = scanner.nextLine();
         String numerosLiterais[] = valores.split(" ");
-        int numeros[] = new int[10];
+        float numeros[] = new float[10];
         for(int i=0;i<10;i++){
-            numeros[i] = Integer.parseInt(numerosLiterais[i]);
+            numeros[i] = Float.parseFloat(numerosLiterais[i]);
         }
         for(int i=9;i>0;i--){
             for(int j=0;j<i;j++){
                 if(numeros[j]>numeros[i]){
-                    int aux = numeros[j];
+                    float aux = numeros[j];
                     numeros[j] = numeros[i];
                     numeros[i] = aux;
                 }
